requirements.txt
# Core Dependencies
gradio==4.44.1
python-dotenv==1.0.0

# Google Cloud Services
google-cloud-documentai==2.20.0
google-cloud-aiplatform==1.38.0
google-cloud-texttospeech==2.15.0
google-cloud-storage==2.14.0

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.0
pdfplumber==0.10.3

# Audio Processing
pydub==0.25.1
librosa==0.10.1

# Web Framework (for API)
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3

# Utilities
requests==2.31.0
numpy==1.24.3
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
